pyproject.toml
[project]
name = "adk-mcp-reddit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.5.0",
    "litellm>=1.69.1",
    "python-dotenv>=1.1.0",
]
